refactored game modals
